// src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const BookingStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  REJECTED: 'REJECTED'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]


export const Role = {
  STUDENT: 'STUDENT',
  TUTOR: 'TUTOR',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  BANNED: 'BANNED',
  PENDING: 'PENDING'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  UNPAID: 'UNPAID'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const PaymentMethod = {
  STRIPE: 'STRIPE',
  SSLCOMMERZ: 'SSLCOMMERZ',
  AMARPAY: 'AMARPAY',
  CASH: 'CASH'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]